// crossnote/src/notebook/types.ts
export type FileSystemStats = {
  mtimeMs: number;
  ctimeMs: number;
  size: number;
  isFile: () => boolean;
  isDirectory: () => boolean;
  isSymbolicLink: () => boolean;
};

export type FileSystemApi = {
  readFile: (path: string, encoding?: BufferEncoding) => Promise<string>;
  writeFile: (
    path: string,
    content: string,
    encoding?: string,
  ) => Promise<void>;
  mkdir: (path: string) => Promise<void>;
  exists: (path: string) => Promise<boolean>;
  stat: (path: string) => Promise<FileSystemStats>;
  readdir: (path: string) => Promise<string[]>;
  unlink: (path: string) => Promise<void>;
};

export type ExtendedMarkdownItOptions = {
  /**
   * Whether to enable the sourceMap
   */
  sourceMap?: boolean;
} & MarkdownIt.Options;

export type MathRenderingOption = 'None' | 'KaTeX' | 'MathJax';

export interface ParserConfig {
  onWillParseMarkdown: (markdown: string) => Promise<string>;
  // NOTE: We disabled this for now
  // opts: { cheerio: CheerioAPI }
  onDidParseMarkdown: (html: string) => Promise<string>;
}